refactor(kubernetes): remove unnecessary outputs from manifest tasks
Tasks only need outputs when they need to share data with tasks in other stages of a pipeline; otherwise, if they are only sharing data with tasks of the same stage, they should only need to use context.

Deck expects `artifacts` and `resolvedExpectedArtifacts` to be present on stage outputs, not context, when rendering the ExecutionArtifactTab, the component that shows artifacts in certain stage's execution details. Otherwise, the Kubernetes-specific stages in Deck are not reading anything off stage outputs to render execution details. Confusingly, the Deploy, Patch, and Run Job (Manifest) stages are all currently reading off the context key `outputs.manifests` (but not from the outputs map itself).